BIGGEST AND BEST SHOW-OFF SCRIPT
The most complicated script I've developed so far. 
It pulls logs from an S3 bucket with varying directories for each CMS server we have, parses the log files for certain lines, pulls the info together into a separate data sheet in a txt file and uploads it to a server directory for QlikView to process and graph, all with LOTS of error-handling and script-setup checks.